requirements.txt
# Web scraping
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.59.8 # o1 support
anthropic>=0.42.0
python-dotenv>=1.0.0

# Testing
unittest2>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

# Google Generative AI
google-generativeai

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.70.0

# SOAP API support for Tradera
zeep>=4.3.0
requests>=2.25.0
